some refactoring
which was neccesary because Status definitions moved to
ControlSystemAdapter
