+++ include/StatusHandler.h
@@ -1,27 +1,14 @@
 #pragma once
 
-#include "DoocsUpdater.h"
-#include "DoocsAdapter.h"
-#include <ChimeraTK/ScalarRegisterAccessor.h>
-#include <ChimeraTK/DataConsistencyGroup.h>
+#include <ChimeraTK/ControlSystemAdapter/StatusWithMessageReaderBase.h>
 #include <boost/shared_ptr.hpp>
-#include <eq_fct.h>
-#include <eq_errors.h>
+#include <boost/noncopyable.hpp>
 
-// TODO discuss - I'm unsure whether pulling in a header from ApplicationCore is ok, if we don't have
-// that dependency in DoocsAdapter otherwise.
-// so maybe, split that header StatusWithMessage.h in a part for DeviceAccess and one for ApplicationCore?
-// But: We use status codes defined in ApplicationCore, StatusAccessorBase::Status
-
-// beware - StatusAccessor name clash with FAULT defined in ArchiveError.h included indirectly by d_fct.h
-#undef FAULT
-#include <ChimeraTK/ApplicationCore/StatusAccessor.h>
-
-#include <ChimeraTK/ApplicationCore/StatusWithMessage.h>
-
-// TODO automatically find status string input, instead of config option.
-
+class EqFct;
 namespace ChimeraTK {
+
+  class DoocsAdapter;
+  class DoocsUpdater;
 
   /**
  * StatusHandler implements an update function for reporting errors to locations.
@@ -35,56 +22,16 @@ namespace ChimeraTK {
     StatusWithMessageReader _varPair;
 
    public:
-    void updateError(TransferElementID transferElementId) {
-      if(!_varPair.update(transferElementId)) return;
-      int statusCode = _varPair._status;
-      int err_no = statusCodeMapping(statusCode);
-      std::string err_str = _varPair.getMessage();
-
-      // Note: we already own the location lock by specification of the DoocsUpdater
-      if(err_no == no_error)
-        _eqFct->set_error(err_no);
-      else
-        _eqFct->set_error(err_no, err_str);
-    }
+    void updateError(TransferElementID transferElementId);
 
     /// statusScalar and statusString are the variables to monitor. statusScalar is mandatory, statusString not.
     /// If both are set, updates must come in consistently
     StatusHandler(EqFct* eqFct, boost::shared_ptr<DoocsUpdater> const& updater,
         boost::shared_ptr<ChimeraTK::NDRegisterAccessor<int32_t>> const& statusScalar,
-        boost::shared_ptr<ChimeraTK::NDRegisterAccessor<std::string>> const& statusString = nullptr)
-    : _doocsUpdater(updater), _eqFct(eqFct), _varPair(ChimeraTK::ScalarRegisterAccessor<int32_t>(statusScalar)) {
-      if(!statusScalar->isReadable()) throw ChimeraTK::logic_error(statusScalar->getName() + " is not readable!");
+        boost::shared_ptr<ChimeraTK::NDRegisterAccessor<std::string>> const& statusString = nullptr);
 
-      _doocsUpdater->addVariable(
-          _varPair._status, _eqFct, std::bind(&StatusHandler::updateError, this, statusScalar->getId()));
-      if(statusString) {
-        if(!statusString->isReadable()) throw ChimeraTK::logic_error(statusString->getName() + " is not readable!");
-        _varPair.setMessageSource(ChimeraTK::ScalarRegisterAccessor<std::string>(statusString));
-
-        _doocsUpdater->addVariable(
-            _varPair._message, _eqFct, std::bind(&StatusHandler::updateError, this, statusString->getId()));
-      }
-    }
-
-    // mapping function from Device.status/StatusOutput to DOOCS error codes
-    int statusCodeMapping(int x) {
-      auto err = StatusAccessorBase::Status(x);
-      // a mapping for StatusOutput values -> DOOCS error codes
-      switch(err) {
-        case StatusAccessorBase::Status::OK:
-          return no_error;
-        case StatusAccessorBase::Status::FAULT:
-          //return ill_function;
-          return not_available; // better compatibility with mapping below
-        case StatusAccessorBase::Status::OFF:
-          return device_offline; // no_error; // or  not_available ?
-        case StatusAccessorBase::Status::WARNING:
-          return warning;
-        default:
-          return x;
-      }
-    }
+    /// mapping function from Device.status/StatusOutput to DOOCS error codes
+    int statusCodeMapping(int x);
   };
 
 } // namespace ChimeraTK
